move pickers to shared component folder
