+++ frontend/src/app/core/components/class-picker/class-picker.component.html
