+++ frontend/src/app/core/components/teacher-picker/teacher-picker.component.spec.ts
